WIP on builder function
