+++ package.json
@@ -66,6 +66,7 @@
         "autoprefixer": "^6.7.2",
         "babel-core": "^6.26.0",
         "babel-loader": "^6.2.10",
+        "babel-plugin-transform-object-rest-spread": "^6.26.0",
         "babel-plugin-transform-react-jsx": "^6.24.1",
         "babel-plugin-transform-runtime": "^6.15.0",
         "babel-preset-env": "^1.6.1",
